backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.0.1
bcrypt==4.1.2
redis==5.0.1
email-validator==2.1.0.post1
Pillow==10.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 